+++ packages/js-client-grpc/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @qdrant/js-client-grpc
 
+## 1.9.0
+
+### Minor Changes
+
+-   Qdrant v1.9.0 API
+
 ## 1.8.0
 
 ### Minor Changes